[0.80.0] C99 sized integers
git-svn-id: svn+ssh://jackaudio.org/trunk/jack@473 0c269be4-1314-0410-8aa9-9f06e86f4224

+++ example-clients/impulse_grabber.c
@@ -149,9 +149,11 @@ main (int argc, char *argv[])
 	*/
 
 	fs = jack_get_sample_rate(client);
-	response_duration = (int)(fs * duration);
+	response_duration = (unsigned long) (fs * duration);
 	response = malloc(response_duration * sizeof(float));
-	fprintf(stderr, "Grabbing %f seconds (%lu samples) of impulse response\n", duration, response_duration);
+	fprintf(stderr,
+		"Grabbing %f seconds (%lu samples) of impulse response\n",
+		duration, response_duration);
 
 	/* create two ports */
